Challenge #1
I've kept the changes in PlayerControler.cs to a minimum to avoid
introducing unwanted consequences as I don't master this code and
don't have access to someone who could help me ensure I won't affect
behaviour, performances or anything else. In a real work environment, if
I needed to modify the code of another dev to be able to test it, I
would first make sure with them that the changes I intend to make are
acceptable ("first" or during the code review).

Added a new PlayerControllerTests file in which I test the jump
functionnality:
- Ensure we go through the jump state as expected when the player start
  to jump
- Ensure that stoping the jump actually impact velocity, making the
  character decelerate to land
- Ensure that the controls that are expected to make the character jump
  indeed do that (to prevent any unwanted behavior modification
  regarding input system)